New Nagios-style check: check_ns_sanity.sh
Quick hack to make sure all nameservers listed for a domain are actually
serving _something_ for that domain.  Doesn't make sure they're all the
same, or up to date, or whatever, just that they all serve something.  We
had an extended badtime as a result of not checking this (and having faith
in our vendor to DTRT).  What better answer than "yep, we're now testing
that"?